changing directory for dictionaries
In many cases you would use your own dictionary models, have a
different way of ‘pathing’ to directory or just a different path. Now
in that case we need a way to change the $this->dataFolder and reload
the rest of the data (assuming with same prefixes, for now). This is a
change to make that possible without requiring to change anything in
working code of those who will maybe merge.